confirm closing unsaved file

+++ notebook.py
@@ -133,9 +133,10 @@ class Notebook(aui.AuiNotebook):
         if index is None: index = self.GetSelection()
         if index >= 0:
             window = self.get_window(index)
-            self.recent_path(window.file_path)
-            self.DeletePage(index)
-            wx.PostEvent(self, NotebookEvent(EVT_NOTEBOOK_TAB_CLOSED, self))
+            if window.confirm_close(self.GetParent(), True):
+                self.recent_path(window.file_path)
+                self.DeletePage(index)
+                wx.PostEvent(self, NotebookEvent(EVT_NOTEBOOK_TAB_CLOSED, self))
         if self.GetPageCount() == 0 and create_untitled:
             self.create_tab()
         self.Thaw()
